Move SendMessageViewModel
